% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_estimation.R
\name{calibrate_traffic_dodgr}
\alias{calibrate_traffic_dodgr}
\title{Calibrate dodgr centrality model}
\usage{
calibrate_traffic_dodgr(
  net_c,
  counts_sf,
  count_col = "avg_daily_car",
  formula = as.formula(paste(count_col, "~ highway + log1p(centrality)"))
)
}
\arguments{
\item{net_c}{Centrality-enriched network (from estimate_traffic_dodgr)}

\item{counts_sf}{sf object with observed traffic counts}

\item{count_col}{Name of the column with traffic counts}

\item{formula}{Model formula for calibration}
}
\description{
Calibrate dodgr centrality model
}
