% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{classify_cycleways}
\alias{classify_cycleways}
\title{Segregation levels}
\usage{
classify_cycleways(cycleways, min_distance = 10)
}
\arguments{
\item{cycleways}{The input dataset for which segregation levels need to be calculated.}

\item{min_distance}{The minimum distance to the road for a cycleway to be considered off-road.}
}
\value{
A an sf object with the new column \code{cycle_segregation} that contains the segregation levels.
}
\description{
This function calculates the segregation levels for a given dataset.
}
