% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{classify_shared_use}
\alias{classify_shared_use}
\title{Classify ways by level of pedestrian/cyclist sharing}
\usage{
classify_shared_use(osm)
}
\arguments{
\item{osm}{An sf object with the road network}
}
\value{
An sf object with the classified cycle network
}
\description{
Ways on which bicycles and pedestrians share space are classified as "Shared use".
According to
}
\details{
tagging includes:
\itemize{
\item highway=path (signposted foot and bicycle path, no dividing line)
\itemize{
\item foot=designated
\item bicycle=designated
\item segregated=no
}
\item highway=path (Signposted foot and bicycle path with dividing line.)
\itemize{
\item segregated=yes
}
\item highway=pedestrian (A way intended for pedestrians)
}
}
\examples{
osm = osm_edinburgh
cycle_network = get_cycling_network(osm)
cycle_network_shared = classify_shared_use(cycle_network)
table(cycle_network_shared$cycle_pedestrian_separation)
plot(cycle_network_shared["cycle_pedestrian_separation"])
# interactive map:
# mapview::mapview(cycle_network_shared, zcol = "cycle_pedestrian_separation")
}
