% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{classify_speeds}
\alias{classify_speeds}
\title{Classify Speeds}
\usage{
classify_speeds(speed_mph)
}
\arguments{
\item{speed_mph}{A numeric vector representing speeds in miles per hour.}
}
\value{
A character vector with the speed categories.
}
\description{
This function classifies speeds in miles per hour (mph) into categories.
}
\examples{
classify_speeds(c(15, 25, 35, 45, 55, 65))
# Returns: "<20 mph", "20 mph", "30 mph", "40 mph", "50 mph", "60+ mph"
}
