% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{clean_speeds}
\alias{clean_speeds}
\title{Clean speeds}
\usage{
clean_speeds(osm)
}
\arguments{
\item{osm}{An sf object with the road network}
}
\value{
An sf object with the cleaned speed values in the column \code{maxspeed_clean}
}
\description{
Clean speeds
}
\examples{
osm = osm_edinburgh
osm_cleaned = clean_speeds(osm)
# check NAs:
sel_nas = is.na(osm_cleaned$maxspeed_clean)
osm_no_maxspeed = osm_cleaned[sel_nas, c("highway")]
table(osm_no_maxspeed$highway) # Active travel infrastructure has no maxspeed
table(osm_cleaned$maxspeed)
table(osm_cleaned$maxspeed_clean)
plot(osm_cleaned[c("maxspeed", "maxspeed_clean")])
}
