% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{clean_widths}
\alias{clean_widths}
\title{Clean cycleway widths (use est_widths when available and width otherwise)}
\usage{
clean_widths(osm)
}
\arguments{
\item{osm}{An sf object with the road network}
}
\value{
An sf object with the cleaned cycleway widths in the column \code{width}
}
\description{
Clean cycleway widths (use est_widths when available and width otherwise)
}
\examples{
osm = osm_edinburgh
osm$est_width = NA
osm$est_width[1:3] = 2
osm_cleaned = clean_widths(osm)
osm$width
osm_cleaned$width
}
