% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{count_bus_lanes}
\alias{count_bus_lanes}
\title{Count how many bus lanes there are}
\usage{
count_bus_lanes(osm)
}
\arguments{
\item{osm}{An sf object with the road network}
}
\value{
The number of bus lanes
}
\description{
Count how many bus lanes there are
}
\examples{
osm = osm_edinburgh
count_bus_lanes(osm)
}
