% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{cycle_net_f}
\alias{cycle_net_f}
\title{Cycle network for Edinburgh, filtered around Leith Walk}
\format{
An sf data frame
}
\description{
This dataset contains the cycle network for Edinburgh, filtered around Leith Walk.
}
\examples{
head(cycle_net_f)
plot(cycle_net_f)
}
