% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{distance_to_road}
\alias{distance_to_road}
\title{Calculate distance from route network segments to roads}
\usage{
distance_to_road(rnet, roads)
}
\arguments{
\item{rnet}{The route network for which the distance to the road needs to be calculated.}

\item{roads}{The road network to which the distance needs to be calculated.}
}
\value{
An sf object with the new column \code{distance_to_road} that contains the distance to the road.
}
\description{
This function approximates the distance from the route network to the nearest road.
It does this by first computing the \code{sf::st_point_on_surface} of the route network segments
and then calculating the distance to the nearest road using the \code{geos::geos_distance} function.
}
\examples{
osm = osm_edinburgh
cycle_network = get_cycling_network(osm)
driving_network = get_driving_network(osm)
edinburgh_cycle_with_distance = distance_to_road(cycle_network, driving_network)
}
