% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{drive_net_f}
\alias{drive_net_f}
\title{Driving network for Edinburgh, filtered around Leith Walk}
\format{
An sf data frame
}
\description{
This dataset contains the driving network for Edinburgh, filtered around Leith Walk.
}
\examples{
head(drive_net_f)
plot(drive_net_f)
}
