% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_estimation.R
\name{estimate_traffic_dodgr}
\alias{estimate_traffic_dodgr}
\title{Estimate traffic using dodgr centrality}
\usage{
estimate_traffic_dodgr(
  osm,
  wt_profile = "motorcar",
  aggregate_by = "osm_id",
  ...
)
}
\arguments{
\item{osm}{An sf object with the road network}

\item{wt_profile}{The weight profile to use (default: "motorcar")}

\item{aggregate_by}{Column to aggregate centrality by (default: "osm_id")}

\item{...}{Additional arguments passed to dodgr::dodgr_centrality}
}
\description{
Estimate traffic using dodgr centrality
}
