% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{et_active}
\alias{et_active}
\title{This function returns OSM keys that are relevant for active travel}
\usage{
et_active()
}
\description{
This function returns OSM keys that are relevant for active travel
}
