% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_active_travel_network}
\alias{get_active_travel_network}
\title{Get the OSM cycling network}
\usage{
get_active_travel_network(osm)
}
\arguments{
\item{osm}{An OSM network object}
}
\value{
A sf object with the OSM network relevant for active travel
with columns containing information on traffic calming measures,
crossings, cycle infrastructure, and more.
}
\description{
Get the OSM cycling network
}
