% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_bus_routes}
\alias{get_bus_routes}
\title{Function to get multilinestrings representing bus routes}
\usage{
get_bus_routes(
  place,
  query = "SELECT * FROM multilinestrings WHERE route == 'bus'",
  extra_tags = "route",
  ...
)
}
\arguments{
\item{place}{A place name or a bounding box passed to \code{osmextract::oe_get()}}

\item{query}{A query to be passed to \code{osmextract::oe_get()}}

\item{extra_tags}{A vector of extra tags to be included in the OSM extract}

\item{...}{Additional arguments passed to \code{osmextract::oe_get()}}
}
\value{
An sf object with the bus routes
}
\description{
It implements the query
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{[out:json][timeout:25];
relation["route"="bus"](\{\{bbox\}\});
out geom;
}\if{html}{\out{</div>}}

See \href{https://overpass-turbo.eu/s/1Xaf}{overpass-turbo.eu}
for an example of the query in action.
}
\examples{
# r = get_bus_routes("Edinburgh")
# r = get_bus_routes("Isle of Wight")
# plot(r["osm_id"])
}
