% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_cycling_network}
\alias{get_cycling_network}
\title{Get the OSM cycling network}
\usage{
get_cycling_network(
  osm,
  ex_c = exclude_highway_cycling(),
  ex_b = exclude_bicycle_cycling()
)
}
\arguments{
\item{osm}{An OSM network object}

\item{ex_c}{A vector of highway values to exclude}

\item{ex_b}{A vector of bicycle values to exclude}
}
\value{
A sf object with the OSM cycling network
}
\description{
Get the OSM cycling network
}
