% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_maxspeed_uk}
\alias{get_maxspeed_uk}
\title{Function to assign default speed limits}
\usage{
get_maxspeed_uk(osm_data)
}
\arguments{
\item{osm_data}{An sf object with the road network}
}
\value{
An sf object with the default speed limits in the column \code{maxspeed}
}
\description{
Function to assign default speed limits
}
