% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_palette_npt}
\alias{get_palette_npt}
\title{Get the palette for the NPT cycle segregation levels}
\usage{
get_palette_npt()
}
\value{
A palette for the NPT cycle segregation levels
}
\description{
Get the palette for the NPT cycle segregation levels
}
\examples{
cols = get_palette_npt()
jsonlite::toJSON(as.list(cols), pretty = TRUE)
col_labs = c("OffRd", "SegW", "SegN", "Share", "Paint")
barplot(seq_along(cols), col = cols, names.arg = col_labs)
}
