% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_points}
\alias{get_points}
\title{Get the OSM network functions}
\usage{
get_points(place, extra_tags = c("traffic_calming", "crossing"), ...)
}
\arguments{
\item{place}{A place name or a bounding box passed to \code{osmextract::oe_get()}}

\item{extra_tags}{A vector of extra tags to be included in the OSM extract}

\item{...}{Additional arguments passed to \code{osmextract::oe_get()}}
}
\value{
A sf object with the OSM network
}
\description{
Get the OSM network functions
}
