% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_traffic_calming}
\alias{get_traffic_calming}
\title{Get the OSM network functions}
\usage{
get_traffic_calming(osm_points, filter = TRUE)
}
\arguments{
\item{osm_points}{An sf object with the OSM network}

\item{filter}{Whether to filter the results to only include traffic calming measures}
}
\value{
A data frame with the OSM network relevant for active travel
}
\description{
Get the OSM network functions
}
\examples{
osm_points = get_points("Edinburgh")
traffic_calming = get_traffic_calming(osm_points)
head(traffic_calming)
plot(traffic_calming["traffic_calming"])
}
