% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{is_wide}
\alias{is_wide}
\title{Classify Separated cycle track by width}
\usage{
is_wide(x, min_width = 2)
}
\arguments{
\item{x}{A numeric vector with the width of the cycleway (m)}

\item{min_width}{The minimum width for a cycleway to be considered wide (m)}
}
\value{
A logical vector indicating whether the cycleway is wide
}
\description{
This function classifies cycleways as wide if the width is greater than or equal to \code{min_width}.
NA values are replaced with 0, meaning that ways with no measurement are considered narrow.
}
\examples{
x = osm_edinburgh$width
x
is_wide(x)
}
