% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{level_of_service}
\alias{level_of_service}
\title{Generate Cycle by Design Level of Service}
\usage{
level_of_service(osm)
}
\arguments{
\item{osm}{An sf object with the road network including speed limits and traffic volumes}
}
\value{
An sf object with the Cycle by Design Level of Service in the column \verb{Level of Service}
}
\description{
Note: you need to have Annual Average Daily Traffic (AADT) values in the dataset
These can be estimated using the \code{estimate_traffic()} function and converted
to CbD AADT categories using the \code{npt_to_cbd_aadt()} function.
}
\examples{
osm = osm_edinburgh
cycle_net = get_cycling_network(osm)             # Get cycle network
driving_network = get_driving_network(osm)       # Get driving network (needed for distance)
cycle_net = distance_to_road(cycle_net, driving_network) # Add distance_to_road column
cycle_net = classify_cycle_infrastructure(cycle_net) # Add cycle_segregation (uses distance)
cycle_net = estimate_traffic(cycle_net)          # Estimate traffic (on cycle net)
cycle_net$AADT = npt_to_cbd_aadt_numeric(cycle_net$assumed_volume) # Add AADT category
# Ensure Speed Limit column exists:
if (!"Speed Limit (mph)" \%in\% names(cycle_net)) {
    cycle_net = clean_speeds(cycle_net)
    cycle_net$`Speed Limit (mph)` = classify_speeds(cycle_net$maxspeed_clean)
}
cycle_net_los = level_of_service(cycle_net)        # Now calculate LoS
table(cycle_net_los$`Level of Service`)
}
