% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{los_table_long}
\alias{los_table_long}
\title{Long format Level of Service (LOS) table}
\format{
A data frame with columns including speed limit, AADT, cycle_segregation and level_of_service
}
\source{
Generated from los_table_npt dataset
}
\usage{
data(los_table_long)
}
\description{
This dataset contains the level of service information in a long format, with columns for speed limit, AADT, infrastructure type, and level of service.
}
\examples{
data(los_table_long)
head(los_table_long)
}
