% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{npt_to_cbd_aadt}
\alias{npt_to_cbd_aadt}
\title{Convert AADT to CBD AADT}
\usage{
npt_to_cbd_aadt(AADT)
}
\arguments{
\item{AADT}{A character or numeric value representing the Annual Average Daily Traffic.}
}
\value{
The converted CBD AADT value.
}
\description{
This function converts Annual Average Daily Traffic (AADT) to Central Business District (CBD) AADT.
It handles both character and numeric inputs by delegating to appropriate helper functions.
}
