% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{npt_to_cbd_aadt_character}
\alias{npt_to_cbd_aadt_character}
\title{Convert AADT categories to CBD AADT character ranges}
\usage{
npt_to_cbd_aadt_character(AADT)
}
\arguments{
\item{AADT}{A character vector representing AADT categories. Valid categories include "0 to 1000", "0 to 2000", "1000+", "All", "1000 to 2000", "2000 to 4000", "2000+", and "4000+".}
}
\value{
A character vector with the converted CBD AADT ranges. Possible return values are "0 to 999", "1000 to 1999", "2000 to 3999", and "4000+".
}
\description{
This function takes an AADT (Annual Average Daily Traffic) category and converts it to a ranges
}
\examples{
npt_to_cbd_aadt_character("0 to 1000") # returns "0 to 999"
npt_to_cbd_aadt_character("1000 to 2000") # returns "1000 to 1999"
npt_to_cbd_aadt_character("2000 to 4000") # returns "2000 to 3999"
npt_to_cbd_aadt_character("4000+") # returns "4000+"
}
