% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{npt_to_cbd_aadt_numeric}
\alias{npt_to_cbd_aadt_numeric}
\title{Convert AADT categories to CBD AADT character ranges}
\usage{
npt_to_cbd_aadt_numeric(AADT)
}
\arguments{
\item{AADT}{A numeric vector representing AADT}
}
\value{
A character vector with the converted CBD AADT ranges. Possible return values are "0 to 999", "1000 to 1999", "2000 to 3999", and "4000+".
}
\description{
This function takes an AADT (Annual Average Daily Traffic) category and converts it to a ranges
}
