% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\docType{data}
\name{osm_edinburgh}
\alias{osm_edinburgh}
\title{Data from edinburgh's OSM network}
\format{
An sf data frame
}
\description{
Data from edinburgh's OSM network
}
\examples{
library(sf)
names(osm_edinburgh)
head(osm_edinburgh)
plot(osm_edinburgh)
}
\keyword{datasets}
