% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\docType{data}
\name{traffic_volumes_edinburgh}
\alias{traffic_volumes_edinburgh}
\alias{traffic_random_edinburgh}
\title{Data from edinburgh's OSM network with traffic volumes}
\format{
A data frame
}
\description{
Data from edinburgh's OSM network with traffic volumes
}
\examples{
head(traffic_volumes_edinburgh)
head(traffic_random_edinburgh)
}
\keyword{datasets}
