% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{classify_cycle_infrastructure}
\alias{classify_cycle_infrastructure}
\title{Segregation levels}
\usage{
classify_cycle_infrastructure(
  osm,
  min_distance = 20,
  classification_type = "Scotland",
  include_mixed_traffic = FALSE
)
}
\arguments{
\item{osm}{The input dataset for which segregation levels need to be calculated.}

\item{min_distance}{The minimum distance to the road for a cycleway to be considered off-road.}

\item{classification_type}{The classification type to be used. Currently only "Scotland" is supported.}

\item{include_mixed_traffic}{Whether to include mixed traffic segments in the results.}
}
\value{
A an sf object with the new column \code{cycle_segregation} that contains the segregation levels.
}
\description{
This function classifies OSM ways in by cycle infrastructure type levels for a given dataset.
}
\details{
See \href{https://wiki.openstreetmap.org/wiki/Key:cycleway}{wiki.openstreetmap.org/wiki/Key:cycleway}
and \href{https://taginfo.openstreetmap.org/keys/cycleway#values}{taginfo.openstreetmap.org/keys/cycleway#values}
for more information on cycleway values used to classify cycle infrastructure.

Currently, only the "Scotland" classification type is supported.
See the Scottish Government's \href{https://www.transport.gov.scot/publication/cycling-by-design/}{Cycling by Design} for more information.
}
\examples{
library(tmap)
tmap_mode("plot")
osm = osm_edinburgh
cycle_network = get_cycling_network(osm)
driving_network = get_driving_network(osm)
netd = distance_to_road(cycle_network, driving_network)
netc = classify_cycle_infrastructure(netd)
library(sf)
plot(netc["cycle_segregation"])
plot(netc["distance_to_road"])
plot_osm_tmap(netc)
# Interactive map:
# tmap_mode("view")
# plot_osm_tmap(netc)
}
