% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_driving_network}
\alias{get_driving_network}
\title{Get the OSM driving network}
\usage{
get_driving_network(osm, ex_d = exclude_highway_driving())
}
\arguments{
\item{osm}{An OSM network object}
}
\value{
A sf object with the OSM driving network
}
\description{
Get the OSM driving network
}
