% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{get_travel_network}
\alias{get_travel_network}
\title{Get the OSM network functions}
\usage{
get_travel_network(place, extra_tags = et_active())
}
\arguments{
\item{place}{A place name or a bounding box passed to \code{osmextract::oe_get()}}

\item{extra_tags}{A vector of extra tags to be included in the OSM extract}
}
\value{
A sf object with the OSM network
}
\description{
Get the OSM network functions
}
