% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{level_of_service}
\alias{level_of_service}
\title{Generate Cycle by Design Level of Service}
\usage{
level_of_service(osm)
}
\arguments{
\item{osm}{An sf object with the road network including speed limits and traffic volumes}
}
\value{
An sf object with the Cycle by Design Level of Service in the column \verb{Level of Service}
}
\description{
Note: you need to have Annual Average Daily Traffic (AADT) values in the dataset
These can be estimated using the \code{estimate_traffic()} function and converted
to CbD AADT categories using the \code{npt_to_cbd_aadt()} function.
}
\examples{
osm = osm_edinburgh
osm = estimate_traffic(osm)
osm$AADT = npt_to_cbd_aadt_numeric(osm$assumed_volume)
osm_los = level_of_service(osm)
}
