% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{level_of_service}
\alias{level_of_service}
\title{Generate Cycle by Design Level of Service}
\usage{
level_of_service(osm)
}
\arguments{
\item{osm}{An sf object with the road network including speed limits and traffic volumes}
}
\value{
An sf object with the Cycle by Design Level of Service in the column \verb{Level of Service}
}
\description{
Generate Cycle by Design Level of Service
}
