% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmactive.R
\name{los_table_complete}
\alias{los_table_complete}
\title{Complete Level of Service (LOS) table}
\format{
A data frame with columns including speed limit, AADT, cycle_segregation and level_of_service
}
\source{
Generated from the classify-cbd vignette
}
\usage{
data(los_table_complete)
}
\description{
This dataset contains the complete level of service information, including missing categories, in a long format.
}
\examples{
data(los_table_complete)
cols = c("Speed Limit (mph)", "Speed (85th kph)")
unique(los_table_complete[cols])
head(los_table_complete)
}
